requirements.txt
Django==2.1.3
django-cors-middleware==1.3.1
django-extensions==2.1.3
djangorestframework==3.9.0
PyJWT==1.6.4
pytz==2018.7
six==1.11.0
